+++ app/src/test/java/dev/pandesal/sbp/extensions/AccountTypeExtensionsTest.kt
@@ -11,6 +11,8 @@ class AccountTypeExtensionsTest {
         assertEquals("Mobile digital wallet", AccountType.MOBILE_DIGITAL_WALLET.label())
         assertEquals("Bank account", AccountType.BANK_ACCOUNT.label())
         assertEquals("Credit card", AccountType.CREDIT_CARD.label())
-        assertEquals("Loan", AccountType.LOAN.label())
+        assertEquals("Loan for asset", AccountType.LOAN_FOR_ASSET.label())
+        assertEquals("Loan for spending", AccountType.LOAN_FOR_SPENDING.label())
     }
 }
+